+++ internal/form/user.go
@@ -7,53 +7,6 @@ import (
 	"gopkg.in/macaron.v1"
 )
 
-//nolint:staticcheck // Reason: needed for legacy code
-type Install struct {
-	DbType   string `binding:"Required"`
-	DbHost   string
-	DbUser   string
-	DbPasswd string
-	DbName   string
-	DbSchema string
-	SSLMode  string
-	DbPath   string
-
-	AppName             string `binding:"Required" locale:"install.app_name"`
-	RepoRootPath        string `binding:"Required"`
-	RunUser             string `binding:"Required"`
-	Domain              string `binding:"Required"`
-	SSHPort             int
-	UseBuiltinSSHServer bool
-	HTTPPort            string `binding:"Required"`
-	AppUrl              string `binding:"Required"`
-	LogRootPath         string `binding:"Required"`
-	EnableConsoleMode   bool
-	DefaultBranch       string
-
-	SMTPHost        string
-	SMTPFrom        string
-	SMTPUser        string `binding:"OmitEmpty;MaxSize(254)" locale:"install.mailer_user"`
-	SMTPPasswd      string
-	RegisterConfirm bool
-	MailNotify      bool
-
-	OfflineMode           bool
-	DisableGravatar       bool
-	EnableFederatedAvatar bool
-	DisableRegistration   bool
-	EnableCaptcha         bool
-	RequireSignInView     bool
-
-	AdminName          string `binding:"OmitEmpty;AlphaDashDot;MaxSize(30)" locale:"install.admin_name"`
-	AdminPasswd        string `binding:"OmitEmpty;MaxSize(255)" locale:"install.admin_password"`
-	AdminConfirmPasswd string
-	AdminEmail         string `binding:"OmitEmpty;MinSize(3);MaxSize(254);Include(@)" locale:"install.admin_email"`
-}
-
-func (f *Install) Validate(ctx *macaron.Context, errs binding.Errors) binding.Errors {
-	return validate(errs, ctx.Data, f, ctx.Locale)
-}
-
 //   __________________________________________.___ _______    ________  _________
 //  /   _____/\_   _____/\__    ___/\__    ___/|   |\      \  /  _____/ /   _____/
 //  \_____  \  |    __)_   |    |     |    |   |   |/   |   \/   \  ___ \_____  \